Export render from logic to index
